pyproject.toml
[project]
name = "cqresearch"
version = "0.1.0"
description = "CryptoQuant Research — BTC/ETH factor-block evolution around the 2024 spot-ETF structural break."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "CryptoQuant Research Team" }]

dependencies = [
    "pandas>=2.2",
    "polars>=1.10",
    "numpy>=1.26",
    "pyarrow>=17.0",
    "statsmodels>=0.14",
    "scikit-learn>=1.5",
    "arch>=7.0",
    "linearmodels>=6.0",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "pypdf>=4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "nbstripout>=0.7",
    "pre-commit>=3.8",
    "ipykernel>=6.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cqresearch"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "scripts", "tools", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
files = ["src/cqresearch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
# Include project root so `from config.paths import ...` resolves, and src/ so
# `import cqresearch` resolves.
pythonpath = [".", "src"]
addopts = "-ra --strict-markers"
